Issue #74: convert internal time format to integer microseconds

+++ specs/src/utils/TimeUtils.h
@@ -7,7 +7,7 @@
 
 #define MICROSECONDS_PER_SECOND 1000000
 
-typedef long double clockValue;
+typedef int64_t clockValue;
 
 clockValue specTimeGetTOD();
